# fix quadratic behavior in natdynlink by using a STW section for frame-descriptor updates