This PR proposes a fix for the quadratic behavior discussed in https://github.com/ocaml/ocaml/pull/11673#issuecomment-1297384255 (which causes performance regressions in both Frama-C and Coq, see also https://github.com/ocaml/ocaml/issues/11926).

We have discussed several implementation strategies to fix the quadratic behavior:

1. use the STW for synchronization: https://github.com/ocaml/ocaml/pull/11673#issuecomment-1399451455
2. use a hashtable with safe concurrent addition: https://github.com/ocaml/ocaml/issues/11926#issuecomment-1399512446
3. use a reader-writer lock: https://github.com/ocaml/ocaml/pull/11673#issuecomment-1399530867

This PR implements option (1): updates to the "current" caml_frame_descrs table are performed in a STW section. This allows unprotected access to caml_frame_descrs from mutators.

Note: I have not tested yet that the quadratic behavior is really gone from Frama-C or Coq. (Compiling them on the current trunk sounds a bit exhausting.) I am thinking of writing a synthetic test instead, but I am not sure how to proceed with that.