@kayceesrk 2023-01-31
> Note: I have not tested yet that the quadratic behavior is really gone from Frama-C or Coq. (Compiling them on the current trunk sounds a bit exhausting.) I am thinking of writing a synthetic test instead, but I am not sure how to proceed with that.

Would it be easier to backport this change to 5.0 and then get Coq and Frama-C to compile? I suspect that the changes between 5.0 and trunk in this part of the code are going to be minimal.