@gasche 2023-01-31
Some performance numbers on a synthetic benchmark, dynlinking 200 .cmxs each with 1000 small functions.

4.14: 1.0s
5.0: 1.7s
trunk: 1.4s
this PR: 1.0s

Good news: this PR does recover the 4.14 performance on this test. (In particular, running 200 stop-the-world has neglectible cost.)

Meh news: the cost of the test appears to be dominated by system-level stuff rather than building the hashtable, so I couldn't observe a nice quadratic plot turning into a linear plot.