@lthls 2023-01-31
I suspect that having 1000 small functions doesn't generate a lot of frametable entries (but I could be wrong).
A single function doing 1000 allocations or external calls might exercise it better (or not, I haven't actually tried).