@stedolan 2023-02-01
> I suspect that having 1000 small functions doesn't generate a lot of frametable entries (but I could be wrong).

This depends on what's in the small functions. Frametable entries are only generated at calls and allocations. In particular, `fun () -> 42` generates no frametable entries at all, but `fun () -> ref 42` generates one.