@gasche 2023-02-01
Each small function in my test contains a function call: the code looks like

```ocaml
let f0 () = 0
let f1 () = f1 () + 1
let f2 () = f2 () + 2
...
```

I could of course tune my synthetic test, but I have the impression that what has been done already answers the question I was looking for: "does this restore 4.14 performance for repeated dynlinking?". The answer is: yes.

I'm thinking of porting a version of the patch (and the buildup PRs) on top of 5.0, so that Frama-C and Coq people can test it. But I think that interested people could already consider reviewing the PR now.