@lthls 2023-02-01
If your code is exactly like that, then inlining will transform all your functions into constants and you will have more or less zero frametable entries. You might want to check the generated Cmm of your code to make sure the calls are still there.