@gadmm 2023-02-01
A benchmark is not necessary to know that the strategy in this PR will result in comparable performance as 4.14 when a single domain is involved. The big unknown is the behaviour in real workloads with multiple domains.

This is the second PR (with Stephen's) that tries to provide a partial fix to the issue, while I have received no indication that the total fix I proposed is something that interests you (and by seeing several partial fixes being worked on, I am led to believe that it actually does not interest you). It might be useful to discuss what is your desired end result before working on specific solutions.