@gasche 2023-02-01
I am not aware of existing programs that load plugins *in parallel* with heavy computations in other domains.

> This is the second PR (with Stephen's) that tries to provide a partial fix to the issue, while I have received no indication that the total fix I proposed is something that interests you (and by seeing several partial fixes being worked on, I am led to believe that it actually does not interest you).

Is this a passive-agressive way to say that I should be working on your proposed solution to the performance regression, rather than my proposed solution to the performance regression? Feel free to send a PR.

Note that an approach that improves scalability could always be proposed later -- it is much easier to do on top of this PR than on top of the trunk as it was before Stephen and myself started working on this.

One advantage of my approach is that it is obvious that there is no cost for programs that only access the frametables, without using Dynlink. With either your approach of the reader-writer mutex proposed by Xavier, we would want to benchmark this to ensure that the synchronization costs for reader do not add overhead. This sounds like work, and I'm too lazy to do it, so I prefer working on the STW approach myself. This is not at all a criticism of both your suggestions, which I think are good ideas as well.

> It might be useful to discuss what is your desired end result before working on specific solutions.

My desired end-result is to solve the performance regression that affects Frama-C and Coq.