@gasche 2023-02-03
I have rebased this PR on top of 5.0 in my remote branch
  https://github.com/gasche/ocaml/commits/frametables-in-stw-5.0

(There are minor conflicts with other changes in trunk, so it is not fully automatic.)

If people interested in Frama-C or Coq would like to try it, they could see whether the dynlink-related slowdown are fixed for their use-case.

cc @maroneze @silene