@gasche 2023-02-03
Note: to test a branch, you probably want to clone the repository+branch and create an opam switch for it, which is described in [HACKING.adoc#opam-switch](https://github.com/ocaml/ocaml/blob/5.0.0/HACKING.adoc#opam-switch).