@gasche 2023-02-06
@lpw25 points out that the need to fix the quadratic behavior may have vanished in practice after Stephen's batching change in #11935. We should consider sending only that change to Frama-C/Coq people to see.