@gasche 2023-02-06
I pushed a 5.0 branch with only the "batch natdynlink calls" change from Stephen:

  https://github.com/gasche/ocaml/tree/natdynlink-batching-5.0

This would also be a candidate to see whether it solves the performance regression in practice. (It is a prefix of the other branch. It does not make the quadratic slowdown go away, but it reduces the number of iterations that give this quadratic behavior by a large constant factor, so it may be enough in practice.)