@maroneze 2023-02-07
Thanks for all the effort and the documentation about compiling a git branch via `opam switch install .`, it's very useful.

I tried cloning your repository and compiling your branch (32e41b9a877d97d3e856cca6d4c5afb62098090e), and ran the OSCS benchmarks, and they seem *extremely* close to those of 4.14.1 (no discernible difference, at least for a first run). Which makes me wonder if I did the right tests.

For instance, the `VERSION` file does contain `5.0.1+dev0-2022-12-15`, but `ocaml --version` in the switch created from `opam switch create . --empty` still outputs `4.14.1`.

Is there something else I could try to make sure I did the right tests?

I'm going to benchmark this branch ([a2c5522](https://github.com/gasche/ocaml/commit/a2c55223899a82179f3bf8a2c265d09d8b55f70d)) and see if I get something different.