@maroneze 2023-02-08
Well, after trying with the other branch, I also got the same times, so I'm definitely running the 4.14 compiler, sorry about the noise.

I just tried erasing some of my switches that were related to `ocaml-variants`, and using the other commands you indicate, and it seems to be actually using the right version this time.

I'll try running the `natdynlink-batching-5.0` branch.

However, for this branch, `frametables-in-stw`, it seems it needs 5.1.0 (instead of 5.0.1), and due to the lack of packages such as `ppx_import`, which are required by Frama-C, I am unable to compile it. Is there a way to compile this branch with a 5.0.1 variant?