@gasche 2023-02-08
Thanks @maroneze for the tests! `frametables-in-stw-5.0` is the version that is backported on top of 5.0.